Titweng_Backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pgvector==0.2.4
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
twilio==8.10.0
reportlab==4.0.7
opencv-python==4.8.1.78
pillow==10.1.0
torch==2.1.1
torchvision==0.16.1
numpy==1.25.2
qrcode[pil]==7.4.2
boto3==1.34.0
secure-smtplib==0.1.1